dev-requirements.py2.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-index --output-file=dev-requirements.py2.txt dev-requirements.in
#
atomicwrites==1.3.0       # via pytest
attrs==19.3.0             # via pytest
backports.functools-lru-cache==1.6.1  # via isort
click==7.1.1              # via pip-tools
configparser==4.0.2       # via flake8, importlib-metadata
contextlib2==0.6.0.post1  # via importlib-metadata
coverage==4.5.4           # via pytest-cov
entrypoints==0.3          # via flake8
enum34==1.1.10            # via flake8
flake8==3.7.9             # via pytest-flake8
funcsigs==1.0.2           # via pytest
functools32==3.2.3.post2  # via flake8
futures==3.3.0            # via isort
importlib-metadata==1.6.0  # via pluggy, pytest
isort==4.3.21             # via pytest-isort
mccabe==0.6.1             # via flake8
more-itertools==5.0.0     # via pytest
packaging==20.3           # via pytest
pathlib2==2.3.5           # via importlib-metadata, pytest
pip-tools==4.5.1          # via -r dev-requirements.in
pluggy==0.13.1            # via pytest
py==1.8.1                 # via pytest
pycodestyle==2.5.0        # via flake8
pyflakes==2.1.1           # via flake8
pyparsing==2.4.7          # via packaging
pytest-ckan==0.0.12       # via -r dev-requirements.in
pytest-cov==2.8.1         # via -r dev-requirements.in
pytest-flake8==1.0.5      # via -r dev-requirements.in
pytest-isort==0.3.1       # via -r dev-requirements.in
pytest==4.6.9             # via -r dev-requirements.in, pytest-ckan, pytest-cov, pytest-flake8, pytest-isort
scandir==1.10.0           # via pathlib2
six==1.14.0               # via more-itertools, packaging, pathlib2, pip-tools, pytest
typing==3.7.4.1           # via flake8
wcwidth==0.1.9            # via pytest
zipp==1.2.0               # via importlib-metadata
mock==3.0.5               # via pytest-ckan
ckantoolkit
